Changing the Gradient
I turned down the opacity on the Gradient a little bit. I think I looks
a lot better and still protects the Image. I'm still not sure that it
looks better than a normal grayscale gradient, but it is a step in the
right direction.
Tell me what you think.