Changed the Gradient Colors to Just Blue
Also made the effect a bit more subtle by changing the opacity again.
Again, I'm not sure of you can revert and then compair which change that
you like better. I would like you to disable and then enable the
gradient to see if you like it better with or without though.